requirements.txt
torch
torchvision
ftfy
regex
git+https://github.com/openai/CLIP.git
Pillow
numpy
matplotlib
mlflow
ninja
